+++ test/compliment/sources/t_class_members.clj
@@ -207,3 +207,28 @@
 
   (fact "static class members have docs"
     (src/static-member-doc "Integer/parseInt" (-ns)) => (checker string?)))
+
+(deftest literals-inference-test
+  (fact "Has around 19 candidates (give or take, varies per JDK),
+which indicates that the members are an exact match against the class of `[]`"
+    (let [c (count (src/members-candidates "." (-ns) (ctx/cache-context
+                                                      "(__prefix__ [])")))]
+      (< 17 c 21))
+    =>
+    truthy)
+
+
+  (fact "A docstring is offered for the previous query"
+    (src/members-doc ".assocN" (-ns)) => (checker string?))
+
+  (fact "Has around 50 candidates (give or take, varies per JDK),
+which indicates that the members are an exact match against the class of `\"\"`"
+    (let [c (count (src/members-candidates "." (-ns) (ctx/cache-context
+                                                      "(__prefix__ \"\")")))]
+      (< 34 c 52))
+    =>
+    truthy)
+
+  (fact "A docstring is offered for the previous query"
+    (src/members-doc ".codePointBefore" (-ns))
+    => (checker string?)))